+++ src/lib/phdi-building-blocks/phdi_building_blocks/linkage.py
@@ -1,6 +1,6 @@
 import hashlib
 import copy
-from phdi_building_blocks.utils import find_patient_resources, get_one_line_address
+from phdi_building_blocks.utils import find_patient_resources, get_one_line_address, get_field_with_use
 
 
 def add_linking_identifier_to_patients(
@@ -10,7 +10,7 @@ def add_linking_identifier_to_patients(
     Given a FHIR resource bundle:
 
     * identify all patient resource(s) in the bundle
-    * extract standardized name, DOB, and address information for each
+    * extract name, DOB, and address information for each
     * compute a unique hash string based on these fields
     * add the hash string to the list of identifiers held in that patient resource
 
@@ -59,29 +59,6 @@ def add_linking_identifier_to_patients(
             }
         )
     return bundle
-
-
-def get_field_with_use(patient: dict, field: str, use: str, default_field: int) -> str:
-    """
-    For a given field (such as name or address), find the first-occuring
-    instance of the field in a given patient JSON dict such that the
-    instance is associated with a particular use. I.e. find the first
-    name for a patient that has an "official" use capacity. If no
-    instance of a field with the requested use case can be found, instead
-    return a specified default field.
-
-    :param patient: Patient from a FHIR bundle
-    :param field: The field to extract
-    :param use: The use the field must have to qualify
-    :param default_field: The index of the field type to treat as
-        the default return type if no field with the requested use case is
-        found
-    :return: The requested use-case-type field
-    """
-    return next(
-        (item for item in patient[field] if item.get("use") == use),
-        patient[field][default_field],
-    )
 
 
 def generate_hash_str(linking_identifier: str, salt_str: str) -> str:
