+++ src/lib/phdi-building-blocks/phdi_building_blocks/utils.py
@@ -10,9 +10,9 @@ def find_patient_resources(bundle: dict) -> List[dict]:
     :return: A list of references to patient dictionaries
     """
     return [
-        r
-        for r in bundle.get("entry")
-        if r.get("resource").get("resourceType") == "Patient"
+        resource
+        for resource in bundle.get("entry")
+        if resource.get("resource").get("resourceType") == "Patient"
     ]
 
 
@@ -29,3 +29,26 @@ def get_one_line_address(address: dict) -> str:
     if "postalCode" in address and address["postalCode"]:
         raw_one_line += f" {address['postalCode']}"
     return raw_one_line
+
+
+def get_field_with_use(patient: dict, field: str, use: str, default_field: int) -> str:
+    """
+    For a given field (such as name or address), find the first-occuring
+    instance of the field in a given patient JSON dict such that the
+    instance is associated with a particular use. I.e. find the first
+    name for a patient that has an "official" use capacity. If no
+    instance of a field with the requested use case can be found, instead
+    return a specified default field.
+
+    :param patient: Patient from a FHIR bundle
+    :param field: The field to extract
+    :param use: The use the field must have to qualify
+    :param default_field: The index of the field type to treat as
+        the default return type if no field with the requested use case is
+        found
+    :return: The requested use-case-type field
+    """
+    return next(
+        (item for item in patient[field] if item.get("use") == use),
+        patient[field][default_field],
+    )
